Cria componente Loader
