+++ src/Auth0/Auth0.callback.js
@@ -1,8 +1,9 @@
 import React, { PureComponent } from 'react';
-import loading from './loading.svg';
+
+import Loader from '../components/Loader/Loader';
 
 class Callback extends PureComponent {
-  componentDidMount () {
+  componentDidMount() {
     const { location, auth } = this.props;
 
     if (/access_token|id_token|error/.test(location.hash)) {
@@ -10,25 +11,8 @@ class Callback extends PureComponent {
     }
   }
 
-  render () {
-    const style = {
-      position: 'absolute',
-      display: 'flex',
-      justifyContent: 'center',
-      height: '100vh',
-      width: '100vw',
-      top: 0,
-      bottom: 0,
-      left: 0,
-      right: 0,
-      backgroundColor: 'white'
-    };
-
-    return (
-      <div style={style}>
-        <img src={loading} alt='loading' />
-      </div>
-    );
+  render() {
+    return <Loader />;
   }
 }
