+++ src/components/Loader/Loader.js
@@ -1,0 +1,24 @@
+import React from 'react';
+
+import loading from './loading.svg';
+
+const style = {
+  position: 'absolute',
+  display: 'flex',
+  justifyContent: 'center',
+  height: '100vh',
+  width: '100vw',
+  top: 0,
+  bottom: 0,
+  left: 0,
+  right: 0,
+  backgroundColor: 'white'
+};
+
+const Loader = () => (
+  <div style={style}>
+    <img src={loading} alt="loading" />
+  </div>
+);
+
+export default Loader;
